+++ gridsome/__tests__/inferTypes.spec.js
@@ -1,4 +1,4 @@
-const inferTypes = require('../lib/graphql/schema/infer-types')
+const { inferTypes } = require('../lib/graphql/schema/infer-types')
 const { GraphQLDate } = require('../lib/graphql/schema/types/date')
 const { fileType } = require('../lib/graphql/schema/types/file')
 const { imageType } = require('../lib/graphql/schema/types/image')
@@ -110,7 +110,7 @@ test('infer image fields', () => {
     }
   ], 'TestPost')
 
-  expect(types.image1.type).toEqual(imageType.type)
+  expect(types.image1.type).toEqual(GraphQLString)
   expect(types.image2.type).toEqual(imageType.type)
   expect(types.image3.type).toEqual(imageType.type)
   expect(types.image4.type).toEqual(imageType.type)